ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.2.16
langchain-groq==0.1.9
langgraph==0.2.28
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
python-json-logger==2.0.7
pytz==2023.3
firebase-admin==6.2.0
